+++ contracts/IMintableBurnableToken.sol
@@ -3,6 +3,7 @@
 pragma solidity ^0.8.0;
 
 interface IMintableBurnableToken {
-  function mint(address to, uint amount) external;
-  function burn(address from, uint amount) external;
+    function mint(address to, uint256 amount) external;
+
+    function burn(address from, uint256 amount) external;
 }
